+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # h.js
 
+
+## h.js 1.0.2  not yet released
+
+* adopt H.matches(start, sel, pat)
+
+
 ## h.js 1.0.0  released 2015/12/15
 
 * initial release